feat(s3): Allow separate logging bucket for AWS Config
For the case where the central logs bucket has Object Lock enabled, AWS
Config does not support logging directly to the central logs bucket. A
workaround for that issue is to create a separate bucket without Object
Lock enabled and replicate everything over to the central logs bucket.
This change adds the option to create that separate bucket with a
replication rule into the central logs bucket and repointing AWS Config
to use that separate bucket instead.